Bytes to KB

+++ db/migrate/20191113061959_add_stats_to_tasks.rb
@@ -1,5 +1,5 @@
 class AddStatsToTasks < ActiveRecord::Migration[5.2]
   def change
-    add_column :tasks, :filesize, :integer
+    add_column :tasks, :filesize, :float
   end
 end
